+++ tests/wasm/main.rs
@@ -41,6 +41,7 @@ pub mod generic_import_args;
 pub mod generic_import_async;
 pub mod generic_import_bounds;
 pub mod generic_import_class_generics;
+pub mod generic_import_closures;
 pub mod generic_import_lifetime;
 pub mod generic_import_methods;
 pub mod generic_import_ref;